refactor: move RawCascadeJob and ExecutionSpecification to domain/run/cascade
These two models are not used anywhere in the blueprint domain — only
in domain/run/* (compile, cascade, service) and the cascade submit test.
Moving them alongside EnvironmentSpecification consumers in run/cascade.py
keeps all execution-related models in the run domain.

domain/blueprint/cascade.py now only defines EnvironmentSpecification,
which is still referenced by BlueprintBuilder (blueprint domain).

Update imports in domain/run/compile.py, domain/run/service.py, and
the two test files (tests/integration/test_submit_job.py and
tests/nonpytest/blueprint.py) to point to the new location.
